requirements.txt
django==3.2
django-crispy-forms
pillow
requests

pylint
pylint-django
